requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy>=2
alembic
pydantic[email]
passlib[bcrypt]
pyjwt
python-dotenv
httpx
icalendar
